Move display core functions into a separate file
